+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   * Added call to record exception with a throwable
   * Improved Countly SDK logging messages
   * Limited the size of crash logs to 10k characters
+  * Limited the size of breadcrumbs to 1000 characters
+  * Limited the amount of breadcrumbs to 1000 entries. If a newer one is added, the oldest one is discarded
   * Fixed a deviceId changing related bug
   * Fixed a bug for setting push consent before init